rename packages to exercise 3
