+++ src/main/java/com/java/master/exercise3/repository/AuthorRepository.java
@@ -1,9 +1,9 @@
-package com.java.master.exercise1.repository;
+package com.java.master.exercise3.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.java.master.exercise1.model.Author;
+import com.java.master.exercise3.model.Author;
 
 @Repository
 public interface AuthorRepository extends JpaRepository<Author, Long> {
