+++ src/main/java/com/java/master/exercise3/service/impl/AuthorServiceImpl.java
@@ -1,13 +1,13 @@
-package com.java.master.exercise1.service.impl;
+package com.java.master.exercise3.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import com.java.master.exercise1.model.Author;
-import com.java.master.exercise1.repository.AuthorRepository;
-import com.java.master.exercise1.service.AuthorService;
+import com.java.master.exercise3.model.Author;
+import com.java.master.exercise3.repository.AuthorRepository;
+import com.java.master.exercise3.service.AuthorService;
 
 @Service
 public class AuthorServiceImpl implements AuthorService {
